requirements.txt
attrs==21.4.0
numpy==1.22.0
